api/requirements.txt
numpy>=1.24.3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
markdown>=3.5.1
scikit-learn>=1.3.0